refactor: components 폴더, types.ts 정리
- components 안에 있는 visualization 관련 components는components/visualization으로 분리
- types.ts안의 type들 sidebar, visualization으로 분리
