+++ frontend/src/types/sidebar.ts
@@ -1,0 +1,7 @@
+export type SidebarSectionProps = {
+    title: string;
+    links: Array<{
+        title: string;
+        path: string;
+    }>;
+};
